Feat: #108 EldersHealthInfoRepository 인터페이스 패턴 리팩토링
- EldersHealthInfoRepository를 인터페이스로 분리
- EldersHealthInfoRepositoryImpl 구현체 생성
- 노인 건강 정보 조회/수정 기능 인터페이스 적용
- 캐싱 기능 포함한 인터페이스 설계
